Добавлена статическая библиотека MathDecimal для подсчета математических операций в Decimal Изменено наследования Все операции переведены в Decimal для точности

+++ MindboxLibrary/FigureCreator.cs
@@ -5,7 +5,7 @@
     /// </summary>
     public class FigureCreator
     {
-        public IFigure CreateFigure(double[] inputValue)
+        public IFigure CreateFigure(decimal[] inputValue)
         {
             switch (inputValue.Length)
             {
